Add python3 pykalman apt package
The upstream pykalman package is satisfied by pip, but in our system, it
installs in opt/ros/noetic/lib/python3.8/site-packages/pykalman instead
of the `python3` directory. There may be another way around this, but
for now, I'm going to try installing it as an apt package. That should
install the package in the normal directory.